backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
psycopg[binary]==3.2.3
python-dotenv==1.0.0
openai==1.30.1
gunicorn==21.2.0
